Engine: reorganized video-related functions
