+++ Engine/ac/global_video.h
@@ -18,7 +18,7 @@
 #ifndef __AGS_EE_AC__GLOBALVIDEO_H
 #define __AGS_EE_AC__GLOBALVIDEO_H
 
-void play_flc_file(int numb,int playflags);
 void scrPlayVideo(const char* name, int skip, int flags);
+void pause_sound_if_necessary_and_play_video(const char *name, int skip, int flags);
 
 #endif // __AGS_EE_AC__GLOBALVIDEO_H
