+++ Engine/media/video/video.h
@@ -18,14 +18,7 @@
 #ifndef __AGS_EE_MEDIA__VIDEO_H
 #define __AGS_EE_MEDIA__VIDEO_H
 
-void calculate_destination_size_maintain_aspect_ratio(int vidWidth, int vidHeight, int *targetWidth, int *targetHeight);
 void play_theora_video(const char *name, int skip, int flags);
-void pause_sound_if_necessary_and_play_video(const char *name, int skip, int flags);
-
-#if defined(WINDOWS_VERSION)
-int __cdecl fli_callback();
-#else
-extern "C" int fli_callback();
-#endif
+void play_flc_file(int numb,int playflags);
 
 #endif // __AGS_EE_MEDIA__VIDEO_H
